+++ packages/compiler/src/core/helpers/index.ts
@@ -1,4 +1,5 @@
 export { DiscriminatedUnion, getDiscriminatedUnion } from "./discriminator-utils.js";
+export { getLocationContext } from "./location-context.js";
 export * from "./operation-utils.js";
 export * from "./path-interpolation.js";
 export * from "./projected-names-utils.js";